Update for new DiamondBank-OG API

+++ src/main/kotlin/net/trueog/questsOG/quests/HomesFive.kt
@@ -137,7 +137,7 @@ class HomesFive : Quest {
     }
 
     override suspend fun consumeQuestItems(player: Player): Boolean {
-        val requiredShards = 2500L * 9
+        val requiredShards = (2500L * 9).toULong()
         val withdrawResult =
             QuestsOG.diamondBankAPI.consumeFromPlayer(
                 player.uniqueId,
